requirements.txt
grpcio==1.57.0
grpcio-tools==1.57.0
protobuf==4.24.0
psycopg2-binary==2.9.7
